Azure Functions/Python/TextAnalyticsAPI/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-ai-textanalytics==5.1.0b3
azure-common==1.1.26
azure-core==1.9.0
certifi==2020.12.5
chardet==3.0.4
idna==2.10
isodate==0.6.0
msrest==0.6.19
oauthlib==3.1.0
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.2
jsonpickle